AsteroidArray and Asteroid classes are split to different source files.
